Refactoring nrcan15_site_term
